Update codebase to upstream revision 884a4e5a35f3aabd804f74b9e6f31a25eda5be7c.

+++ drivers/bluetooth/hci/userchan.c
@@ -181,7 +181,7 @@ static void rx_thread(void *p1, void *p2, void *p3)
 
 	LOG_DBG("started");
 
-	uint16_t frame_size = 0;
+	ssize_t frame_size = 0;
 
 	while (1) {
 		static uint8_t frame[512];
